backend/requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Data validation
pydantic==2.5.3
python-dotenv==1.0.0

# Image processing
Pillow==10.2.0

# PyTorch and CUDA
torch==2.2.0
torchvision==0.17.0

# Hugging Face and Diffusers
transformers==4.40.0
accelerate==0.28.0
diffusers @ git+https://github.com/huggingface/diffusers.git

# Additional ML dependencies
safetensors==0.4.2
sentencepiece==0.2.0

# Utilities
requests==2.31.0
